Don't export FESpaces with module_local=True
Strange segfaults appear when exporting with this setting enabled,
specifically, the FESpace was destructed before a Gridfunction (still
holding a shared_ptr of FESpace) was destructed.

Caveat: Its' not possible to compile the module multiple times,
you need to restart the Jupyter kernel in case you change the source
code.